feat context에 있던 내용 가져왔습니다~! context의 provider에 하나 추가했습니다. 제대로 동작하는 거 확인했습니다~! #2
